adding opengl triangle

+++ engine/input/InputManager.h
@@ -10,8 +10,8 @@ public:
     void update();
     [[nodiscard]] bool isQuit() const;
     [[nodiscard]] bool isKeyDown(SDL_Scancode key) const;
-    [[nodiscard]] int getMoveX() const;
-    [[nodiscard]] int getMoveY() const;
+    [[nodiscard]] float getMoveX() const;
+    [[nodiscard]] float getMoveY() const;
 private:
     bool quit = false;
     const Uint8* keyboardState = nullptr;
